refactor: moved helpers to a separate file
